refactor: TrasactionModal component
