+++ src/components/NewTrasactionModal/styles.ts
@@ -1,0 +1,3 @@
+import styled from 'styled-components';
+
+export const Heading = styled.h2``;
